Clean up pkg dependencies list
The 'pkg install' fails when there's a duplicate entry on the list of packages
to install, e.g:

  pkg: No packages matching 'curl' has been found in the repositories

This happens when either there's explicit duplicate or a package gets on the
list indirectly because of a dependency on it.

This commit works around this problem by avoiding duplicates, but we should
fix the root cause in 'pkg' one day.