nova: Suppress using dhcp_release upon instace deletion
The dhcp_release is part of dnsmasq, but does not get built in FreeBSD port
because of the Linux-specific netlink dependency.  As a workaround we suppress
forcing dhcp_release upon instance delete and rely on the system to reclaim
the unused IPs upon lease expiry (together with a short timeout period).